コンポーネント化した
