+++ src/components/atoms/Icon.js
@@ -1,0 +1,17 @@
+import { IconButton } from "@chakra-ui/button";
+
+
+export const Icon = (props) => {
+   
+    return(
+        <IconButton
+        ml="auto"
+        area-aria-label="メニューボタン"
+        icon={props.icon}
+        size="md"
+        variant="unStyled"
+        display={{ base: "block", md: "none" }}
+        onClick={props.click}
+      />
+    )
+}
